kafka integration added error event creating and stats counting, refactoring http integration
